Changed: DC_engine, unit_test Removed: DC_engine::LOAD_K_2 and unnecessary comments in the code
Next objective: To add inverse part of the exponental part of the load.
Reason: To enable opportunity to implement the static and non static
friction force moment as the load.

For example:
DC_engine::LOAD_K_0 as a static friction force
and there are needed to be exp part that can decrease smoothly
this value while velocity is increasing to imitate the change of
the static friction force to the non static one.